requirements.txt
Flask
python-docx
PyMuPDF
reportlab
pypandoc